Extract configuration parsing to its own crate and use configuration objects as parameters, adding in some extra configuration in anticipation of being used later.

+++ gravibumpers/visual_representation/src/lib.rs
@@ -42,7 +42,7 @@ pub trait SequenceAnimator {
         particle_map_sequence: impl std::iter::ExactSizeIterator<
             Item = impl std::iter::ExactSizeIterator<Item = impl data_structure::ParticleRepresentation>,
         >,
-        milliseconds_per_frame: u32,
+        milliseconds_per_frame: u16,
         output_filename: &str,
     ) -> Result<(), Box<dyn std::error::Error>>;
 }
